Size

+++ 0x00-hello_world/6-size.c
@@ -9,10 +9,10 @@ char c;
 int i;
 float f;
  printf("Size of a char: %lu.\n", (unsigned char)sizeof(c));
-printf("Size of an int: %lu.\n", (unsigned int)sizeof (i));
+printf("Size of an int: %lu.\n", (unsigned short int)sizeof (i));
 printf("Size of a long int: %ln.\n", (unsigned long int)sizeof (i));
 printf("Size of a long long int: %ln.\n", (unsigned long long int)sizeof (i));
-printf("Size of a float: %ln.\n", (unsigned float)sizeof (f));
+ printf("Size of a float: %ln.\n", (float)sizeof (f));
 return (0);
 }
